feat: Read GITHUB_TOKEN if token is not passed
Reads environment variable GITHUB_TOKEN if the token is not passed
via the `--token` parameter. If neither are passed will proceed
unauthenticated.

Using the envvar GITHUB_TOKEN to match what is available when run
from inside an action in GitHub. So that auth will automatically
just work if this is run from a GitHub Workflow.

Closes issue #19.

Ref: https://docs.github.com/en/actions/security-for-github-actions/security-guides/automatic-token-authentication
Co-Authored-By: Mikaël Barbero <mikael.barbero@gmail.com>
Signed-off-by: Thanh Ha <zxiiro@gmail.com>